+++ lib/onlyoffice_documentserver_testing_framework/test_instance_docs/doc_test_site_functions.rb
@@ -232,6 +232,7 @@ class DocTestSiteFunctions
   def self.supported_languages(version = OnlyofficeDocumentserverTestingFramework::DocumentServerVersion.new(8, 1, 0))
     version_class = OnlyofficeDocumentserverTestingFramework::DocumentServerVersion
     version_mappings = [
+      [version_class.new(8, 2), 'doc_test_site_languages_after_8_2.list'],
       [version_class.new(8, 1), 'doc_test_site_languages_after_8_1.list'],
       [version_class.new(8, 0), 'doc_test_site_languages_after_8_0.list'],
       [version_class.new(7, 4), 'doc_test_site_languages_after_7_4.list'],