+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * Add language list for DocumentServer v8.1
+* Add language list for DocumentServer v8.2
 * Simplify `DocumentServerVersion#supported_languages` method
 * Fix `rubocop-1.65.0` cop `Gemspec/AddRuntimeDependency`